[Feat] 탐색 컨텐츠 기초 제작 완료, 유저데이터 테이블 수정, 로딩 씬 제작 완료

+++ Assets/TIV/Script/MVVM/ShipMenuUViewModelExtension.cs
@@ -53,7 +53,7 @@ namespace ViewModel.Extensions
             vm.EquipedReactorKey = shipData.GetEquipedReactor();
             vm.EquipedRadiatorKey = shipData.GetEquipedRadiator();
         }
-        public static void OnChangeUserItemData(this ShipMenuUIViewModel vm, long credit, long superCredit)//콜백
+        public static void OnChangeUserItemData(this ShipMenuUIViewModel vm, long credit, int superCredit)//콜백
         {
             vm.Credit = credit;
             vm.SuperCredit = superCredit;
